Auth Schema Service some checks add

+++ src/main/java/elbar/company/learn_center_rest/service/auth/block/AuthBlockServiceImpl.java
@@ -11,14 +11,17 @@ import elbar.company.learn_center_rest.repository.auth.block.AuthBlockRepository
 import elbar.company.learn_center_rest.response.Data;
 import elbar.company.learn_center_rest.service.AbstractService;
 import elbar.company.learn_center_rest.validator.auth.block.AuthBlockValidator;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.webjars.NotFoundException;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -38,15 +41,25 @@ public class AuthBlockServiceImpl extends AbstractService<AuthBlockValidator, Au
     @Override
     public ResponseEntity<Data<Void>> update(AuthBlockUpdateDTO DTO) {
         validator.validOnUpdate(DTO);
-        AuthBlock authBlock = repository.getByCode(DTO.getCode());
+        Optional<AuthBlock> optional = repository.getByCode(DTO.getCode());
+        if (optional.isEmpty()) {
+            throw new NotFoundException("Auth block not found");
+        }
+        AuthBlock authBlock = optional.get();
         authBlock.setDuration(DTO.getDuration());
         authBlock.setBlockedReason(DTO.getBlockedReason());
         authBlock.setUpdatedAt(LocalDateTime.now());
+        repository.save(authBlock);
         return new ResponseEntity<>(new Data<>(true), HttpStatus.OK);
     }
 
     @Override
     public ResponseEntity<Data<Void>> delete(UUID key) {
+        validator.validateKey(key);
+        Optional<AuthBlock> optional = repository.getByCode(key);
+        if (optional.isEmpty()) {
+            throw new NotFoundException("Auth block not found");
+        }
         repository.deleteByCode(key);
         return new ResponseEntity<>(new Data<>(true), HttpStatus.OK);
     }
@@ -54,13 +67,13 @@ public class AuthBlockServiceImpl extends AbstractService<AuthBlockValidator, Au
     @Override
     public ResponseEntity<Data<AuthBlockGetDTO>> get(UUID key) {
         validator.validateKey(key);
-        return new ResponseEntity<>(new Data<>(mapper.fromGetDTO(repository.getByCode(key))), HttpStatus.OK);
+        return new ResponseEntity<>(new Data<>(mapper.fromGetDTO(repository.getByCode(key).orElseThrow(() -> new NotFoundException("Auth block not found")))), HttpStatus.OK);
     }
 
     @Override
     public ResponseEntity<Data<AuthBlockDetailDTO>> detail(UUID key) {
         validator.validateKey(key);
-        return new ResponseEntity<>(new Data<>(mapper.fromDetailDTO(repository.getByCode(key))), HttpStatus.OK);
+        return new ResponseEntity<>(new Data<>(mapper.fromDetailDTO(repository.getByCode(key).orElseThrow(() -> new NotFoundException("Auth block not found")))), HttpStatus.OK);
     }
 
     @Override
