+++ src/main/java/elbar/company/learn_center_rest/service/auth/card_type/CardTypeServiceImpl.java
@@ -16,9 +16,11 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.webjars.NotFoundException;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -37,7 +39,11 @@ public class CardTypeServiceImpl extends AbstractService<CardTypeValidator, Card
     @Override
     public ResponseEntity<Data<Void>> update(CardTypeUpdateDTO DTO) {
         validator.validOnUpdate(DTO);
-        AuthCardType cardType = repository.getByCode(DTO.getCode());
+        Optional<AuthCardType> optional = repository.getByCode(DTO.getCode());
+        if (optional.isEmpty()) {
+            throw new NotFoundException("Card Type not found");
+        }
+        AuthCardType cardType = optional.get();
         cardType.setName(DTO.getName());
         cardType.setPublished(DTO.isPublished());
         cardType.setUpdatedAt(LocalDateTime.now());
@@ -48,6 +54,10 @@ public class CardTypeServiceImpl extends AbstractService<CardTypeValidator, Card
     @Override
     public ResponseEntity<Data<Void>> delete(UUID key) {
         validator.validateKey(key);
+        Optional<AuthCardType> optional = repository.getByCode(key);
+        if (optional.isEmpty()) {
+            throw new NotFoundException("Card Type not found");
+        }
         repository.deleteByCode(key);
         return new ResponseEntity<>(new Data<>(true), HttpStatus.OK);
     }
@@ -55,13 +65,13 @@ public class CardTypeServiceImpl extends AbstractService<CardTypeValidator, Card
     @Override
     public ResponseEntity<Data<CardTypeGetDTO>> get(UUID key) {
         validator.validateKey(key);
-        return new ResponseEntity<>(new Data<>(mapper.fromGetDTO(repository.getByCode(key))), HttpStatus.OK);
+        return new ResponseEntity<>(new Data<>(mapper.fromGetDTO(repository.getByCode(key).orElseThrow(() -> new NotFoundException("Card Type not found")))), HttpStatus.OK);
     }
 
     @Override
     public ResponseEntity<Data<CardTypeDetailDTO>> detail(UUID key) {
         validator.validateKey(key);
-        return new ResponseEntity<>(new Data<>(mapper.fromDetailDTO(repository.getByCode(key))), HttpStatus.OK);
+        return new ResponseEntity<>(new Data<>(mapper.fromDetailDTO(repository.getByCode(key).orElseThrow(() -> new NotFoundException("Card Type not found")))), HttpStatus.OK);
     }
 
     @Override
