+++ src/main/java/elbar/company/learn_center_rest/service/auth/payment/PaymentServiceImpl.java
@@ -17,8 +17,10 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.webjars.NotFoundException;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -42,6 +44,10 @@ public class PaymentServiceImpl extends AbstractService<PaymentValidator, Paymen
     @Override
     public ResponseEntity<Data<Void>> delete(UUID key) {
         validator.validateKey(key);
+        Optional<AuthPayment> optional = repository.getByCode(key);
+        if (optional.isEmpty()) {
+            throw new NotFoundException("Language not found");
+        }
         repository.deleteByCode(key);
         return new ResponseEntity<>(new Data<>(true), HttpStatus.OK);
     }
@@ -49,13 +55,13 @@ public class PaymentServiceImpl extends AbstractService<PaymentValidator, Paymen
     @Override
     public ResponseEntity<Data<PaymentGetDTO>> get(UUID key) {
         validator.validateKey(key);
-        return new ResponseEntity<>(new Data<>(mapper.fromGetDTO(repository.getByCode(key))), HttpStatus.OK);
+        return new ResponseEntity<>(new Data<>(mapper.fromGetDTO(repository.getByCode(key).orElseThrow(() -> new NotFoundException("Payment not found")))), HttpStatus.OK);
     }
 
     @Override
     public ResponseEntity<Data<PaymentDetailDTO>> detail(UUID key) {
         validator.validateKey(key);
-        return new ResponseEntity<>(new Data<>(mapper.fromDetailDTO(repository.getByCode(key))), HttpStatus.OK);
+        return new ResponseEntity<>(new Data<>(mapper.fromDetailDTO(repository.getByCode(key).orElseThrow(() -> new NotFoundException("Payment not found")))), HttpStatus.OK);
     }
 
     @Override
