+++ src/main/java/elbar/company/learn_center_rest/service/auth/role/RoleServiceImpl.java
@@ -16,8 +16,10 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.webjars.NotFoundException;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -40,6 +42,10 @@ public class RoleServiceImpl extends AbstractService<RoleValidator, RoleMapper, 
     @Override
     public ResponseEntity<Data<Void>> delete(UUID key) {
         validator.validateKey(key);
+        Optional<Role> optional = repository.getByCode(key);
+        if (optional.isEmpty()) {
+            throw new NotFoundException("Role not found");
+        }
         repository.deleteByCode(key);
         return new ResponseEntity<>(new Data<>(true), HttpStatus.OK);
     }
@@ -47,13 +53,13 @@ public class RoleServiceImpl extends AbstractService<RoleValidator, RoleMapper, 
     @Override
     public ResponseEntity<Data<RoleGetDTO>> get(UUID key) {
         validator.validateKey(key);
-        return new ResponseEntity<>(new Data<>(mapper.fromGetDTO(repository.getByCode(key))), HttpStatus.OK);
+        return new ResponseEntity<>(new Data<>(mapper.fromGetDTO(repository.getByCode(key).orElseThrow(() -> new NotFoundException("Role not found")))), HttpStatus.OK);
     }
 
     @Override
     public ResponseEntity<Data<RoleDetailDTO>> detail(UUID key) {
         validator.validateKey(key);
-        return new ResponseEntity<>(new Data<>(mapper.fromDetailDTO(repository.getByCode(key))), HttpStatus.OK);
+        return new ResponseEntity<>(new Data<>(mapper.fromDetailDTO(repository.getByCode(key).orElseThrow(() -> new NotFoundException("Role not found")))), HttpStatus.OK);
     }
 
     @Override
