+++ src/main/java/elbar/company/learn_center_rest/service/auth/status/StatusServiceImpl.java
@@ -17,14 +17,15 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.webjars.NotFoundException;
 
 import javax.transaction.Transactional;
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 @Service
-@Transactional
 public class StatusServiceImpl extends AbstractService<StatusValidator, StatusMapper, StatusRepository> implements StatusService {
     public StatusServiceImpl(StatusValidator validator, StatusMapper mapper, StatusRepository repository) {
         super(validator, mapper, repository);
@@ -40,7 +41,11 @@ public class StatusServiceImpl extends AbstractService<StatusValidator, StatusMa
     @Override
     public ResponseEntity<Data<Void>> update(StatusUpdateDTO DTO) {
         validator.validOnUpdate(DTO);
-        Status status = repository.getByCode(DTO.getCode());
+        Optional<Status> optional = repository.getByCode(DTO.getCode());
+        if (optional.isEmpty()) {
+            throw new NotFoundException("Status not found");
+        }
+        Status status = optional.get();
         status.setName(DTO.getName());
         status.setPublished(DTO.getIsPublished());
         status.setUpdatedAt(LocalDateTime.now());
@@ -51,6 +56,10 @@ public class StatusServiceImpl extends AbstractService<StatusValidator, StatusMa
     @Override
     public ResponseEntity<Data<Void>> delete(UUID key) {
         validator.validateKey(key);
+        Optional<Status> optional = repository.getByCode(key);
+        if (optional.isEmpty()) {
+            throw new NotFoundException("Status not found");
+        }
         repository.deleteByCode(key);
         return new ResponseEntity<>(new Data<>(true), HttpStatus.OK);
     }
@@ -58,13 +67,13 @@ public class StatusServiceImpl extends AbstractService<StatusValidator, StatusMa
     @Override
     public ResponseEntity<Data<StatusGetDTO>> get(UUID key) {
         validator.validateKey(key);
-        return new ResponseEntity<>(new Data<>(mapper.fromGetDTO(repository.getByCode(key))), HttpStatus.OK);
+        return new ResponseEntity<>(new Data<>(mapper.fromGetDTO(repository.getByCode(key).orElseThrow(() -> new NotFoundException("Status not found")))), HttpStatus.OK);
     }
 
     @Override
     public ResponseEntity<Data<StatusDetailDTO>> detail(UUID key) {
         validator.validateKey(key);
-        return new ResponseEntity<>(new Data<>(mapper.fromDetailDTO(repository.getByCode(key))), HttpStatus.OK);
+        return new ResponseEntity<>(new Data<>(mapper.fromDetailDTO(repository.getByCode(key).orElseThrow(() -> new NotFoundException("Status not found")))), HttpStatus.OK);
     }
 
     @Override
