+++ src/main/java/elbar/company/learn_center_rest/service/auth/teacher/TeacherServiceImpl.java
@@ -16,14 +16,15 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.webjars.NotFoundException;
 
 import javax.transaction.Transactional;
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 @Service
-@Transactional
 public class TeacherServiceImpl extends AbstractService<TeacherValidator, TeacherMapper, TeacherRepository> implements TeacherService {
     public TeacherServiceImpl(TeacherValidator validator, TeacherMapper mapper, TeacherRepository repository) {
         super(validator, mapper, repository);
@@ -39,7 +40,11 @@ public class TeacherServiceImpl extends AbstractService<TeacherValidator, Teache
     @Override
     public ResponseEntity<Data<Void>> update(TeacherUpdateDTO DTO) {
         validator.validOnUpdate(DTO);
-        Teacher teacher = repository.getByCode(DTO.getCode());
+        Optional<Teacher> optional = repository.getByCode(DTO.getCode());
+        if (optional.isEmpty()) {
+            throw new NotFoundException("Teacher not found");
+        }
+        Teacher teacher = optional.get();
         teacher.setAbout(DTO.getAbout());
         teacher.setUpdatedAt(LocalDateTime.now());
         teacher.setAge(DTO.getAge());
@@ -51,6 +56,10 @@ public class TeacherServiceImpl extends AbstractService<TeacherValidator, Teache
     @Override
     public ResponseEntity<Data<Void>> delete(UUID key) {
         validator.validateKey(key);
+        Optional<Teacher> optional = repository.getByCode(key);
+        if (optional.isEmpty()) {
+            throw new NotFoundException("Teacher not found");
+        }
         repository.deleteByCode(key);
         return new ResponseEntity<>(new Data<>(true), HttpStatus.OK);
     }
@@ -58,13 +67,13 @@ public class TeacherServiceImpl extends AbstractService<TeacherValidator, Teache
     @Override
     public ResponseEntity<Data<TeacherGetDTO>> get(UUID key) {
         validator.validateKey(key);
-        return new ResponseEntity<>(new Data<>(mapper.fromGetDTO(repository.getByCode(key))), HttpStatus.OK);
+        return new ResponseEntity<>(new Data<>(mapper.fromGetDTO(repository.getByCode(key).orElseThrow(() -> new NotFoundException("Teacher not found")))), HttpStatus.OK);
     }
 
     @Override
     public ResponseEntity<Data<TeacherDetailDTO>> detail(UUID key) {
         validator.validateKey(key);
-        return new ResponseEntity<>(new Data<>(mapper.fromDetailDTO(repository.getByCode(key))), HttpStatus.OK);
+        return new ResponseEntity<>(new Data<>(mapper.fromDetailDTO(repository.getByCode(key).orElseThrow(() -> new NotFoundException("Teacher not found")))), HttpStatus.OK);
     }
 
     @Override
