+++ src/main/java/elbar/company/learn_center_rest/service/auth/user_role/UserRoleServiceImpl.java
@@ -16,8 +16,10 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.webjars.NotFoundException;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -41,6 +43,10 @@ public class UserRoleServiceImpl extends AbstractService<UserRoleValidator, User
     @Override
     public ResponseEntity<Data<Void>> delete(UUID key) {
         validator.validateKey(key);
+        Optional<AuthUserRole> optional = repository.getByCode(key);
+        if (optional.isEmpty()) {
+            throw new NotFoundException("Auth User Role not found");
+        }
         repository.deleteByCode(key);
         return new ResponseEntity<>(new Data<>(true), HttpStatus.OK);
     }
@@ -48,13 +54,13 @@ public class UserRoleServiceImpl extends AbstractService<UserRoleValidator, User
     @Override
     public ResponseEntity<Data<UserRoleGetDTO>> get(UUID key) {
         validator.validateKey(key);
-        return new ResponseEntity<>(new Data<>(mapper.fromGetDTO(repository.getByCode(key))), HttpStatus.OK);
+        return new ResponseEntity<>(new Data<>(mapper.fromGetDTO(repository.getByCode(key).orElseThrow(() -> new NotFoundException("Auth User Role not found")))), HttpStatus.OK);
     }
 
     @Override
     public ResponseEntity<Data<UserRoleDetailDTO>> detail(UUID key) {
         validator.validateKey(key);
-        return new ResponseEntity<>(new Data<>(mapper.fromDetailDTO(repository.getByCode(key))), HttpStatus.OK);
+        return new ResponseEntity<>(new Data<>(mapper.fromDetailDTO(repository.getByCode(key).orElseThrow(() -> new NotFoundException("Auth User Role not found")))), HttpStatus.OK);
     }
 
     @Override
